Updated BMS segment & created BMS Microcontroller
Updated BMS segment schematic to to reflect that the LTC6804 is powered
from the cells it is monitoring.  Also added connectors.  Created a
preliminary design for the microcontroller board that will process the
BMS data from all the BMS segment boards.  Some resistor values and pin
assignments still need to be determined.